backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
python-decouple==3.8
psycopg2-binary==2.9.6
djoser==2.2.0
django-extra-fields==3.0.2
Pillow==9.5.0
django-filter==23.2
django-cors-headers==4.0.0